requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
jinja2>=3.1.2
playwright>=1.40.0
google-generativeai>=0.3.0
tenacity>=8.2.3
pydantic>=2.5.0
pytest>=7.4.3
pytest-mock>=3.12.0
colorama>=0.4.6
tqdm>=4.66.1
